pyproject.toml
[project]
name = "bluesky-nats"
version = "0.1.1"
description = "NATS JetStream connectivity for Bluesky"
readme = "README.md"
authors = [
    {name = "Niko Kivel", email = "niko.kivel@lightsource.ca"},
]
requires-python = ">=3.10"
dependencies = [
    "bluesky>=1.12.0",
    "nats-py>=2.9.0",
    "ormsgpack>=1.5.0",
    "ruff>=0.6.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
